+++ lib/src/cli/task.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 - 2022 Buijs Software
+// Copyright (c) 2021 - 2023 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -52,7 +52,7 @@ abstract class Task {
   List<Task> dependsOn() => [];
 
   /// Output list of example commands to use this Task.
-  List<String> exampleCommands() => [];
+  List<String> exampleCommands();
 
   @override
   String toString() {