+++ test/src/cli/task_service_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 - 2022 Buijs Software
+// Copyright (c) 2021 - 2023 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -246,6 +246,9 @@ class _DummyTask extends Task {
 
   @override
   Future<void> toBeExecuted(String pathToRoot) async {}
+
+  @override
+  List<String> exampleCommands() => [];
 }
 
 class _SelfDependingTask extends _DummyTask {