+++ test/src/common/utilities_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 - 2022 Buijs Software
+// Copyright (c) 2021 - 2023 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -31,4 +31,9 @@ void main() {
             e is KlutterException &&
             e.cause.startsWith("Path does not exist"))));
   });
+
+  test("Verify _substitute correctly normalizes a path", () {
+    expect(File("foo/bar/res/../../pikachu").normalizeToFile.path.endsWith("foo/pikachu"), true);
+  });
+
 }