+++ test/src/producer/android_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 - 2022 Buijs Software
+// Copyright (c) 2021 - 2023 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -184,6 +184,17 @@ void main() {
             e is KlutterException &&
             e.cause.startsWith("Path does not exist:") &&
             e.cause.endsWith("fake"))));
+  });
+
+  test("Verify exception is thrown if root/android/gradle  does not exist", () {
+    final android = Directory("${Directory.systemTemp.path}${s}wag1")
+      ..createSync(recursive: true);
+    expect(
+            () =>
+                setGradleWrapperVersion(pathToAndroid: android.absolutePath),
+        throwsA(predicate((e) =>
+        e is KlutterException &&
+            e.cause.contains("Missing gradle-wrapper.properties file in"))));
   });
 
   test(